% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advert.R
\docType{data}
\name{advert}
\alias{advert}
\title{Sales and advertising expenditure}
\format{
Data frame containing the following columns:
\describe{
  \item{advert}{Monthly Advertising expenditure}
  \item{sales}{Monthly sales volume}
}
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 6.7.
Exercise 8.1.
}
\usage{
advert
}
\description{
Monthly sales and advertising expenditure for an automotive parts company.
}
\examples{
plot(sales ~ advert, data=advert)

}
\keyword{datasets}
