% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advsales.R
\docType{data}
\name{advsales}
\alias{advsales}
\title{Sales volume and advertising expenditure}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 8.
}
\usage{
advsales
}
\description{
Sales volume and advertising expenditure for a dietary weight control
product.
}
\examples{
plot(advsales)
}
\references{
Blattberg and Jeuland (1981).
}
\keyword{datasets}
