\name{AdvStatComp-package}
\alias{AdvStatComp-package}
\alias{AdvStatComp}
\docType{package}
\title{
Package containing functions for my project for the course Advanced Statistical Computing (Fall 2015)
}
\description{
The necessary functions to do the simulation study are shipped with the package. This allows the package to be loaded on workers in parallel processing. 
}
\details{
\tabular{ll}{
Package: \tab AdvStatComp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-11-14\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Sebastian Andersson

Maintainer: Sebastian Andersson <sebastian.andersson@statistik.uu.se>
}
\references{

}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
