% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{Mutual savings bank deposits}
\format{Data frame containing the following columns:
\describe{
  \item{EOM}{End of month balance}
  \item{AAA}{Composite AAA bond rates}
  \item{threefour}{US Government 3-4 year bonds}
}}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 6.
}
\usage{
bank
}
\description{
Deposits in a mutual savings bank in a large metropolitan area.
}
\examples{
plot(bank)

}
\keyword{datasets}
