% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beer.R
\docType{data}
\name{beer}
\alias{beer}
\title{Monthly beer production}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 2.
}
\usage{
beer
}
\description{
Monthly Australian beer production: Jan 1991 -- Aug 1995.
}
\examples{
plot(beer)
seasonplot(beer)
tsdisplay(beer)

}
\keyword{datasets}
