% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicoal.R
\docType{data}
\name{bicoal}
\alias{bicoal}
\title{Annual bituminous coal production}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 7.7.
}
\usage{
bicoal
}
\description{
Annual bituminous coal production in the USA: 1920--1968.
}
\examples{
tsdisplay(bicoal)

}
\keyword{datasets}
