% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston}
\alias{boston}
\title{Monthly dollar volume of sales}
\format{Bivariate time series containing the following columns:
\describe{
  \item{nyase}{New York and American Stock Exchange dollar volume}
  \item{bse}{Boston Stock Exchange dollar volume}
}}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 6.5
}
\usage{
boston
}
\description{
Monthly dollar volume of sales on Boston stock exchange and combined New
York and American stock exchange. January 1967 -- November 1969.
}
\examples{
plot(boston)

}
\references{
McGee and Carleton (1970) Piecewise regression, \emph{Journal of
the American Statistical Association}, \bold{65}, 1109--1124.
}
\keyword{datasets}
