% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricksq.R
\docType{data}
\name{bricksq}
\alias{bricksq}
\title{Quarterly clay brick production}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 1 and
Exercise 2.3.
}
\usage{
bricksq
}
\description{
Australian quarterly clay brick production: 1956--1994.
}
\examples{
plot(bricksq)
seasonplot(bricksq)
tsdisplay(bricksq)

}
\keyword{datasets}
