% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canadian.R
\docType{data}
\name{canadian}
\alias{canadian}
\title{Canadian unemployment rate}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 4.1.
}
\usage{
canadian
}
\description{
Canadian unemployment rate as a percentage of the civilian labor force
between 1974 and the third quarter of 1975.
}
\examples{
plot(canadian)

}
\keyword{datasets}
