% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capital}
\alias{capital}
\title{Quarterly capital expenditure and appropriations}
\format{Bivariate time series containing the following columns:
\describe{
  \item{capital}{Quarterly capital expenditure for US manufacturing.}
  \item{appropriations}{Quarterly capital appropriations for US manufacturing.}
}}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 8.
}
\usage{
capital
}
\description{
Seasonally adjusted quarterly capital expenditure and appropriations in U.S.
manufacturing: 1953--1974.
}
\examples{
plot(capital)

}
\keyword{datasets}
