\name{Check.Correct.Model}
\alias{Check.Correct.Model}
\title{Check correct model}
\description{
  Very simple wrapper for any().
}
\usage{
Check.Correct.Model(input.model, target.model)
}
\arguments{
  \item{input.model}{The input model.}
  \item{target.model}{The target model.}
}
\keyword{Check.Correct.Model}