% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chicken.R
\docType{data}
\name{chicken}
\alias{chicken}
\title{Price of chicken}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 9.
}
\usage{
chicken
}
\description{
Price of chicken in US (constant dollars): 1924--1993.
}
\examples{
plot(chicken)


}
\keyword{datasets}
