% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condmilk.R
\docType{data}
\name{condmilk}
\alias{condmilk}
\title{Condensed milk}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 7.5.
}
\usage{
condmilk
}
\description{
Manufacturer's Stocks of evaporated and sweetened condensed milk.
}
\examples{
plot(condmilk)
seasonplot(condmilk)
tsdisplay(condmilk)
}
\keyword{datasets}
