% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper.R
\docType{data}
\name{copper}
\alias{copper}
\title{Copper price}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 9.
}
\usage{
copper
}
\description{
Yearly copper prices, 1800--1997 (in constant 1997 dollars).
}
\examples{
plot(copper)

}
\keyword{datasets}
