% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper1.R
\docType{data}
\name{copper1}
\alias{copper1}
\title{Copper prices}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 9.
}
\usage{
copper1
}
\description{
Monthly copper prices for 28 consecutive months (in constant 1997 dollars).
}
\examples{
plot(copper1)
}
\keyword{datasets}
