% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper2.R
\docType{data}
\name{copper2}
\alias{copper2}
\title{Copper prices}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 9.
}
\usage{
copper2
}
\description{
Yearly copper prices for 14 consecutive years (in constant 1997 dollars).
}
\examples{
plot(copper2)
}
\keyword{datasets}
