% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper3.R
\docType{data}
\name{copper3}
\alias{copper3}
\title{Copper prices}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 9.
}
\usage{
copper3
}
\description{
Yearly copper prices for 43 consecutive years (in constant 1997 dollars).
}
\examples{
plot(copper3)
}
\keyword{datasets}
