\name{cpimel}
\alias{cpimel}
\title{Consumer price index}
\description{Quarterly CPI (consumer price index) for Victoria: Q1 1980 to Q2 1995.}
\usage{cpimel}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 8.7.}

\keyword{datasets}
\examples{tsdisplay(cpimel)
}
