% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpimel.R
\docType{data}
\name{cpimel}
\alias{cpimel}
\title{Consumer price index}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 8.7.
}
\usage{
cpimel
}
\description{
Quarterly CPI (consumer price index) for Victoria: Q1 1980 to Q2 1995.
}
\examples{
tsdisplay(cpimel)
}
\keyword{datasets}
