\name{Create.FMA}
\alias{Create.FMA}
\title{Create.FMA}
\description{
  Create list for frequensti model averaging estimators.
}
\usage{
Create.FMA(w, name, allModels)
}
\arguments{
  \item{w}{The weight vector.}
  \item{name}{The name of the method.}
  \item{allModels}{The resulting list from \link{\code{EstAllModels}}.}
}
\keyword{Create.FMA}