\name{Data.Generation}
\alias{Data.Generation}
\title{Generate data}
\description{
  Generates data. 
}
\usage{
Data.Generation(samplesize, covariance, error.sd, mu, theta.x, col.index, j.x, j.z, 
new.samplesize)
}
\arguments{
  \item{samplesize}{The sample size.}
  \item{covariance}{The covariance matrix for the independent variables.}
  \item{error.sd}{The error standard deviation.}
  \item{mu}{The mean of the response.}
  \item{theta.x}{The coefficients.}
  \item{col.index}{The column index.}
  \item{j.x}{The number of Xs.}
  \item{j.z}{The number of Zs.}
}
\keyword{Data.Generation}