% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dj.R
\docType{data}
\name{dj}
\alias{dj}
\title{Dow-Jones index}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 7.
}
\usage{
dj
}
\description{
Dow-Jones index on 251 trading days ending 26 Aug 1994.
}
\examples{
tsdisplay(dj)

}
\references{
Brockwell and Davis (1996)
}
\keyword{datasets}
