% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dole.R
\docType{data}
\name{dole}
\alias{dole}
\title{Unemployment benefits in Australia}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 2.3.
}
\usage{
dole
}
\description{
Monthly total of people on unemployment benefits in Australia (Jan 1965 --
Jul 1992).
}
\examples{
plot(dole)
tsdisplay(dole)

}
\keyword{datasets}
