% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowjones.R
\docType{data}
\name{dowjones}
\alias{dowjones}
\title{Dow-Jones index}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 2.7.
}
\usage{
dowjones
}
\description{
Dow-Jones index, 28 Aug - 18 Dec 1972.
}
\examples{
tsdisplay(dowjones)
}
\keyword{datasets}
