% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{econsumption}
\alias{econsumption}
\title{Electricity consumption and temperature}
\format{Data frame containing the following columns:
\describe{
  \item{Mwh}{Daily electricity consumption (megawatt-hours)}
  \item{temp}{Daily maximum temperature (degrees Celsius)}
}}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 5.5
}
\usage{
econsumption
}
\description{
Electricity consumption and maximum temperature for 12 randomly chosen days.
}
\examples{
plot(Mwh ~ temp, data=econsumption, pch=19, col=4)

}
\keyword{datasets}
