% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs.R
\docType{data}
\name{eggs}
\alias{eggs}
\title{Price of eggs}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 9.
}
\usage{
eggs
}
\description{
Price of dozen eggs in US, 1900--1993, in constant dollars.
}
\examples{
plot(eggs)
}
\keyword{datasets}
