% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eknives.R
\docType{data}
\name{eknives}
\alias{eknives}
\title{Sales of electric knives}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 4.2.
}
\usage{
eknives
}
\description{
Sales of electric knives: Jan 1991 - April 1992.
}
\examples{
plot(eknives)
}
\keyword{datasets}
