% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elco.R
\docType{data}
\name{elco}
\alias{elco}
\title{Sales of Elco's laser printers}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 10.
}
\usage{
elco
}
\description{
Sales of Elco's laser printers: 1992--1998.
}
\examples{
plot(elco)
}
\keyword{datasets}
