% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elec.R
\docType{data}
\name{elec}
\alias{elec}
\title{Electricity production}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapters 1--2, 7.
}
\usage{
elec
}
\description{
Australian monthly electricity production: Jan 1956 -- Aug 1995.
}
\examples{
plot(elec)
seasonplot(elec)
tsdisplay(elec)

}
\keyword{datasets}
