% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expenditure.R
\docType{data}
\name{expenditure}
\alias{expenditure}
\title{Expenditure}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 2.
}
\usage{
expenditure
}
\description{
Expenditure for 12 supermarket customers.
}
\examples{
hist(expenditure)
}
\keyword{datasets}
