% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy.R
\docType{data}
\name{fancy}
\alias{fancy}
\title{Sales for a souvenir shop}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 5.8.
}
\usage{
fancy
}
\description{
Monthly sales for a souvenir shop on the wharf at a beach resort town in
Queensland, Australia.
}
\examples{
plot(fancy)
seasonplot(fancy)

}
\keyword{datasets}
