% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fma-package.R
\docType{package}
\name{fma-package}
\alias{fma-package}
\alias{fma}
\title{Data sets from "Forecasting: methods and applications" by Makridakis,
Wheelwright and Hyndman (1998)}
\description{
All data sets from "Forecasting: methods and applications" by Makridakis,
Wheelwright and Hyndman (Wiley, 3rd ed., 1998).
}
\references{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York.
\url{https://robjhyndman.com/forecasting/}
}
\author{
Rob J Hyndman. \email{Rob.Hyndman@monash.edu}
}
\keyword{package}
