% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/french.R
\docType{data}
\name{french}
\alias{french}
\title{Industry index}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 4.4.
}
\usage{
french
}
\description{
French index of industry.
}
\examples{
plot(french)
}
\keyword{datasets}
