\name{fsales}
\alias{fsales}
\title{reports}
\description{Quarterly reports of a French company.}
\usage{fsales}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 3.7.}

\keyword{datasets}
\examples{plot(fsales)
seasonplot(fsales)
}
