\name{Generate.S}
\alias{Generate.S}
\title{Generate the S matrix}
\description{
  Generation of the S matrix. 
}
\usage{
Generate.S(Fix.X, Fix.Z, Potential.X, Potential.Z)
}
\arguments{
  \item{Fix.X}{Number of fixed Xs.}
  \item{Fix.Z}{Number of fixed Zs.}
  \item{Potential.X}{Vector of indexes for potential Xs.}
  \item{Potential.Z}{Vector of indexes for potential Zs.}
}
\keyword{Generate.S}