% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsales.R
\docType{data}
\name{hsales}
\alias{hsales}
\title{Sales of one-family houses}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 3.
}
\usage{
hsales
}
\description{
Monthly sales of new one-family houses sold in the USA since 1973.
}
\examples{
plot(hsales)
plot(stl(hsales,"periodic"),main="Sales of new one-family houses, USA")

}
\references{
US Census Bureau, Manufacturing and Construction Division
}
\keyword{datasets}
