% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsales2.R
\docType{data}
\name{hsales2}
\alias{hsales2}
\title{Sales of new one-family houses}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 7.10.
}
\usage{
hsales2
}
\description{
Sales of new one-family houses in the USA (Jan 1987 -- Nov 1995).
}
\examples{
plot(hsales2)
seasonplot(hsales2)
tsdisplay(hsales2)
}
\keyword{datasets}
