% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huron.R
\docType{data}
\name{huron}
\alias{huron}
\title{Level of Lake Huron}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 8.2.
}
\usage{
huron
}
\description{
Level of Lake Huron in feet (reduced by 570 feet): 1875--1972.
}
\examples{
plot(huron)

}
\keyword{datasets}
