% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibmclose.R
\docType{data}
\name{ibmclose}
\alias{ibmclose}
\title{Closing IBM stock price}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 7.2.
}
\usage{
ibmclose
}
\description{
Daily closing IBM stock price.
}
\examples{
tsdisplay(ibmclose)

}
\references{
Box, Jenkins and Reinsell (1994) \emph{Time series analysis:
forecasting and control}, 3rd edition, Holden-Day: San Francisco.
}
\keyword{datasets}
