% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{data}
\name{input}
\alias{input}
\title{Input series}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 8.6.
}
\usage{
input
}
\description{
Input series for exercise 8.6.
}
\examples{
plot(input)
}
\keyword{datasets}
