% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internet.R
\docType{data}
\name{internet}
\alias{internet}
\title{Number of internet users}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 7.
}
\usage{
internet
}
\description{
Number of users logged on to an internet server each minute over a
100-minute period.
}
\examples{
tsdisplay(internet)

}
\keyword{datasets}
