% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invent15.R
\docType{data}
\name{invent15}
\alias{invent15}
\title{Inventory demand}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 2.6. Also
Chapter 4.
}
\usage{
invent15
}
\description{
Inventory demand for product E15.
}
\examples{
plot(invent15)
}
\keyword{datasets}
