% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kkong}
\alias{kkong}
\title{King Kong data}
\format{Data frame consisting of following columns
\describe{
  \item{weight}{Weights of 21 gorillas}
  \item{height}{Heights of 21 gorillas}
}}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 5. Exercise
5.6.
}
\usage{
kkong
}
\description{
King Kong data.
}
\examples{
plot(weight~height, data=kkong, pch=19, col=2)

}
\keyword{datasets}
