% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mink.R
\docType{data}
\name{mink}
\alias{mink}
\title{Number of minks trapped}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 2.4.
}
\usage{
mink
}
\description{
Annual number of minks trapped in McKenzie river district of northwest
Canada: 1848--1911.
}
\examples{
tsdisplay(mink)
}
\keyword{datasets}
