% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mma.R
\name{mma}
\alias{mma}
\title{MMA}
\usage{
mma(X, formula = NULL, ycol = 1, variance = c("BA", "boot"), bsa = 200)
}
\arguments{
\item{X}{dataframe including X,Y}

\item{formula}{linear model}

\item{ycol}{1}

\item{variance}{"BA", "boot"}

\item{bsa}{number of bootstrap}
}
\value{
coefficients  averaging.weights  setup  coef
}
\description{
MMA
}
\examples{

}
