% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motel}
\alias{motel}
\title{Total accommodation at hotel, motel and guest house}
\format{Trivariate time series containing the following columns:
  \describe{
    \item{Roomnights}{Total room nights}
    \item{Takings}{Total monthly takings (thousands of dollars)}
    \item{CPI}{Quarterly CPI values}
}}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 8.7.
}
\usage{
motel
}
\description{
Total room nights occupied and total monthly takings from accommodation at
hotel, motel and guest house in Victoria, Australia: Jan 1980 - June 1995.
}
\examples{
plot(motel[,2],motel[,1], xlab="Room nights", ylab="Takings",pch=19,col=4)

}
\keyword{datasets}
