\name{motion}
\alias{motion}
\title{Employment figures in the motion picture industry}
\description{Monthly employment figures for the motion picture industry (SIC Code 78): Jan 1955 -- Dec 1970.}
\usage{motion}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 7.9.}

\references{"Employment and earnings, US 1909--1978", Department
of Labor, 1979.}

\keyword{datasets}
\examples{plot(motion)
seasonplot(motion)
tsdisplay(motion)
}
