% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nail.R
\docType{data}
\name{nail}
\alias{nail}
\title{Nail prices}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 9.
}
\usage{
nail
}
\description{
Nail prices, 1800--1996 in constant dollars.
}
\examples{
plot(nail)
}
\keyword{datasets}
