% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oilprice.R
\docType{data}
\name{oilprice}
\alias{oilprice}
\title{Oil prices}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 10.
}
\usage{
oilprice
}
\description{
Oil prices in constant 1997 dollars: 1870--1997.
}
\examples{
plot(oilprice)

}
\keyword{datasets}
