% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ozone}
\alias{ozone}
\title{Ozone depletion and melanoma rates}
\format{Data frame containing the following columns:
\describe{
  \item{ozonedep}{Ozone depletion rates as percentages}
  \item{melanoma}{Melanoma rates as percentages}
}}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 5.3.
}
\usage{
ozone
}
\description{
Ozone depletion and melanoma rates in various locations.
}
\examples{
plot(ozonedep~melanoma,data=ozone,pch=19,col=2)

}
\keyword{datasets}
