% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paris.R
\docType{data}
\name{paris}
\alias{paris}
\title{Average temperature}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 2.1.
}
\usage{
paris
}
\description{
Average monthly temperature in Paris.
}
\examples{
plot(paris)
seasonplot(paris)
tsdisplay(paris)
}
\keyword{datasets}
