% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcv}
\alias{pcv}
\title{GDP}
\format{Bivariate time series consisting of the following columns
\describe{
  \item{GDP}{GDP Western Europe}
  \item{PCV}{PCV Industry sales}
}}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 5.
}
\usage{
pcv
}
\description{
GDP for Western Europe and PCV industry sales.
}
\examples{
plot(PCV~GDP,data=pcv,pch=20,col=2)

}
\keyword{datasets}
