% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petrol.R
\docType{data}
\name{petrol}
\alias{petrol}
\title{Sales of petroleum and related product}
\format{Multivariate time series data:
\describe{
  \item{Chemicals}{Sales of chemicals and allied products}
  \item{Coal}{Sales of Bituminous coal products}
  \item{Petrol}{Sales of petroleum and coal products}
  \item{Vehicles}{Sales of motor vehicles and parts}
}}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 8.
}
\usage{
petrol
}
\description{
US monthly sales of petroleum and related product: Jan 1971 - Dec 1991.
}
\examples{
plot(petrol)
}
\keyword{datasets}
