% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigs.R
\docType{data}
\name{pigs}
\alias{pigs}
\title{Number of pigs slaughtered}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 7.
}
\usage{
pigs
}
\description{
Monthly total number of pigs slaughtered in Victoria, Australia (Jan 1980 --
Aug 1995).
}
\examples{
tsdisplay(pigs)
}
\keyword{datasets}
