% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plastics.R
\docType{data}
\name{plastics}
\alias{plastics}
\title{Sales of plastic product}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 3.5.
}
\usage{
plastics
}
\description{
Monthly sales of product A for a plastics manufacturer.
}
\examples{
plot(plastics)
seasonplot(plastics)
plot(stl(plastics,"periodic"))

}
\keyword{datasets}
