% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollution.R
\docType{data}
\name{pollution}
\alias{pollution}
\title{Shipment of pollution equipment}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 7.
}
\usage{
pollution
}
\description{
Monthly shipments of pollution equipment (in thousands of French francs),
Jan 1986 -- Oct 1996.
}
\examples{
tsdisplay(pollution)

}
\keyword{datasets}
