% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productC.R
\docType{data}
\name{productC}
\alias{productC}
\title{Sales of product C}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 1.
}
\usage{
productC
}
\description{
Sales of product C (a lubricant sold in large containers).
}
\examples{
plot(productC)

}
\keyword{datasets}
