% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qelec.R
\docType{data}
\name{qelec}
\alias{qelec}
\title{Electricity production}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 3.4.
}
\usage{
qelec
}
\description{
Quarterly electricity production.
}
\examples{
plot(decompose(qelec))

}
\keyword{datasets}
