% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsales.R
\docType{data}
\name{qsales}
\alias{qsales}
\title{Sales data}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 3.7 and
Table 4-7.
}
\usage{
qsales
}
\description{
Quarterly exports of a French company in thousands of francs.
}
\examples{
plot(qsales)

}
\keyword{datasets}
