% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{running}
\alias{running}
\title{Running times and maximal aerobic capacity}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 2.5.
}
\usage{
running
}
\description{
Running times and maximal aerobic capacity for 14 female runners.
}
\examples{
plot(times~capacity, data=running, pch=19, col=2)

}
\references{
Conley, Krahenbuhl, Burkett and Millar (1981) Physiological
correlates of female road racing performance, \emph{Research Quarterly
Exercise Sport}, \bold{52}, 441--448.
}
\keyword{datasets}
