% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\docType{data}
\name{sales}
\alias{sales}
\title{Sales data}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 5.
}
\usage{
sales
}
\description{
Sales data over 10 time periods.
}
\examples{
plot(sales,type="p")
abline(lsfit(1:10,sales))

}
\keyword{datasets}
