% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schizo.R
\docType{data}
\name{schizo}
\alias{schizo}
\title{Perceptual speed scores}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 8.8.
}
\usage{
schizo
}
\description{
Daily perceptual speed scores for a schizophrenic patient. The patient began
receiving a powerful tranquilizer (chlorpromzaine) on the 61st day and
continued receiving the drug for the remainder of the sample period. It is
expected that this drug would reduce perceptual speed.
}
\examples{
plot(schizo)

}
\references{
McCleary and Hay (1980).
}
\keyword{datasets}
