% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shampoo.R
\docType{data}
\name{shampoo}
\alias{shampoo}
\title{Sales of shampoo}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 3.
}
\usage{
shampoo
}
\description{
Sales of shampoo over a three year period.
}
\examples{
plot(shampoo)

}
\keyword{datasets}
