% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheep.R
\docType{data}
\name{sheep}
\alias{sheep}
\title{Sheep population}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 7.6.
}
\usage{
sheep
}
\description{
Sheep population (in millions) of England and Wales: 1867--1939.
}
\examples{
tsdisplay(sheep)

}
\references{
Kendall (1976).
}
\keyword{datasets}
