% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ship.R
\docType{data}
\name{ship}
\alias{ship}
\title{Electric can opener shipments}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 4. Exercise
4.6.
}
\usage{
ship
}
\description{
Electric can opener shipments.
}
\examples{
plot(ship)

}
\keyword{datasets}
