% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipex.R
\docType{data}
\name{shipex}
\alias{shipex}
\title{Shipments}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 3.1
}
\usage{
shipex
}
\description{
Shipments
}
\examples{
plot(shipex)

}
\keyword{datasets}
