\name{Sim.Iter}
\alias{Sim.Iter}
\title{Simulation function}
\description{
  A function which corresponds to one iteration in the simulation.
}
\usage{
Sim.Iter(estimators = Estimators, samplesize = SampleSize, covariance = Covariance, 
error.sd = Error.SD, mu = Mu, theta.x = Theta.X, j.x = J.X, j.z = J.Z, 
col.index = Col.Index, new.samplesize = New.SampleSize, 
include.intercept = Include.Intercept, s.matrix = S.matrix, 
correct.model.form = Correct.Model.Form, potential.gamma = Potential.Gamma, 
poppars = popPars, fix.x = Fix.X, fix.z = Fix.Z, potential.x = Potential.X, 
potential.z = Potential.Z)
}
\arguments{
}
\keyword{Sim.Iter}