% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strikes.R
\docType{data}
\name{strikes}
\alias{strikes}
\title{Number of strikes}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 7.4
}
\usage{
strikes
}
\description{
Number of strikes in the US from 1951 to 1980.
}
\examples{
tsdisplay(strikes)
}
\references{
Brockwell and Davis (1991)
}
\keyword{datasets}
