% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telephone.R
\docType{data}
\name{telephone}
\alias{telephone}
\title{Telephone cost}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 9.
}
\usage{
telephone
}
\description{
Telephone cost in San Francisco, New York: 1915--1996.
}
\examples{
plot(telephone)

}
\keyword{datasets}
