% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{texasgas}
\alias{texasgas}
\title{Price and consumption of natural gas}
\format{Data frame containing the following columns:
\describe{
  \item{price}{Average price in cents per thousand cubic feet}
  \item{consumption}{Consumption per customer in thousand cubic feet.}
}}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 5.10.
Exercise 6.2.
}
\usage{
texasgas
}
\description{
Price and per capita consumption of natural gas in 20 towns in Texas.
}
\examples{
plot(consumption ~ price, data=texasgas)

}
\keyword{datasets}
