% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukdeaths.R
\docType{data}
\name{ukdeaths}
\alias{ukdeaths}
\title{Total deaths and serious injuries}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 8.
}
\usage{
ukdeaths
}
\description{
Monthly total deaths and serious injuries on UK roads: Jan 1975 -- Dec 1984.
In February 1983, new legislation came into force requiring seat belts to be
worn.
}
\examples{
plot(ukdeaths)
seasonplot(ukdeaths)
tsdisplay(ukdeaths)

}
\references{
Harvey (1989)
}
\keyword{datasets}
