% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdeaths.R
\docType{data}
\name{usdeaths}
\alias{usdeaths}
\title{Accidental deaths in USA}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercises 2.3 and
2.4.
}
\usage{
usdeaths
}
\description{
Monthly accidental deaths in USA.
}
\examples{
plot(usdeaths)
seasonplot(usdeaths)
tsdisplay(usdeaths)

}
\keyword{datasets}
