% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uselec.R
\docType{data}
\name{uselec}
\alias{uselec}
\title{Total generation of electricity}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 7.8.
}
\usage{
uselec
}
\description{
Monthly total generation of electricity by the U.S. electric industry (Jan
1985 - Oct 1996.
}
\examples{
plot(uselec)
seasonplot(uselec)
tsdisplay(uselec)

}
\keyword{datasets}
