% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustreas.R
\docType{data}
\name{ustreas}
\alias{ustreas}
\title{Treasury bill contracts}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 1.
}
\usage{
ustreas
}
\description{
US treasury bill contracts on the Chicago market for 100 consecutive trading
days in 1981.
}
\examples{
plot(ustreas)
tsdisplay(ustreas)

}
\keyword{datasets}
