% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagesuk.R
\docType{data}
\name{wagesuk}
\alias{wagesuk}
\title{Real daily wages}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 9.
}
\usage{
wagesuk
}
\description{
Real daily wages in pound, England: 1260--1994.
}
\examples{
plot(wagesuk)
}
\keyword{datasets}
