% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheat.R
\docType{data}
\name{wheat}
\alias{wheat}
\title{Wheat prices}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 9.
}
\usage{
wheat
}
\description{
Wheat prices in constant 1996 pounds: 1264--1996.
}
\examples{
plot(wheat)

}
\keyword{datasets}
