% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wn.R
\docType{data}
\name{wn}
\alias{wn}
\title{White noise series}
\format{Time series data}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Exercise 7.3.
}
\usage{
wn
}
\description{
White noise series.
}
\examples{
tsdisplay(wn)
}
\keyword{datasets}
