% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnoise.R
\docType{data}
\name{wnoise}
\alias{wnoise}
\title{White noise time series}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chpater 7.
}
\usage{
wnoise
}
\description{
White noise time series with 36 values.
}
\examples{
tsdisplay(wnoise)

}
\keyword{datasets}
