% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\docType{data}
\name{writing}
\alias{writing}
\title{Sales of printing and writing paper}
\format{
Time series data
}
\source{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting:
methods and applications}, John Wiley & Sons: New York. Chapter 7.
}
\usage{
writing
}
\description{
Industry sales for printing and writing paper (in thousands of French
francs): Jan 1963 -- Dec 1972.
}
\examples{
tsdisplay(writing)
seasonplot(writing)

}
\keyword{datasets}
