\name{advsales}
\alias{advsales}
\title{Sales volume and advertising expenditure}
\description{Sales volume and advertising expenditure for a dietary weight control product.}
\usage{advsales}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley
& Sons: New York. Chapter 8.}
\references{Blattberg and Jeuland (1981).}
\keyword{datasets}
\examples{plot(advsales)
}
