\name{AdvStatComp-package}
\alias{AdvStatComp-package}
\alias{AdvStatComp}
\docType{package}
\title{
Package containing functions for my project for the course Advanced Statistical Computing (Fall 2015)
}
\description{
The necessary functions to do the simulation study are shipped with the package. This allows the package to be loaded on workers in parallel processing. 
}
\details{
\tabular{ll}{
Package: \tab AdvStatComp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-11-14\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Sebastian Andersson

Maintainer: Sebastian Andersson <sebastian.andersson@statistik.uu.se>
}
\references{
Hansen, Bruce (2007) Least Squares Model Averaging, Econometrica, 75(4):1175-1189.
}

\keyword{ package }
