\name{beer}
\alias{beer}
\title{Monthly beer production}
\description{Monthly Australian beer production: Jan 1991 -- Aug 1995.}
\usage{beer}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley
& Sons: New York. Chapter 2.}

\keyword{datasets}
\examples{plot(beer)
seasonplot(beer)
tsdisplay(beer)
}
