\name{bicoal}
\alias{bicoal}
\title{Annual bituminous coal production}
\description{Annual bituminous coal production in the USA: 1920--1968.}
\usage{bicoal}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley
& Sons: New York. Exercise 7.7.}

\keyword{datasets}
\examples{tsdisplay(bicoal)
}
