\name{books}
\alias{books}
\title{Sales of paperback and hardcover books}
\description{Daily sales of paperback and hardcover books at the same store.}

\usage{books}
\format{Bivariate time series containing the following columns:
\describe{\item{Paperback}{Number of paperback sales each day}
\item{Hardcover}{Number of hardcover sales each day}
}}

\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley
& Sons: New York. Exercise 4.5.}

\keyword{datasets}

\examples{plot(books)
}
