\name{bricksq}
\alias{bricksq}
\title{Quarterly clay brick production}
\description{Australian quarterly clay brick production: 1956--1994.}
\usage{bricksq}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley
& Sons: New York. Chapter 1 and Exercise 2.3.}

\keyword{datasets}
\examples{plot(bricksq)
seasonplot(bricksq)
tsdisplay(bricksq)
}
