\name{canadian}
\alias{canadian}
\title{Canadian unemployment rate}
\description{Canadian unemployment rate as a percentage of
the civilian labor force between 1974 and the third quarter of 1975.}
\usage{canadian}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley
& Sons: New York. Exercise 4.1.}

\keyword{datasets}
\examples{plot(canadian)
}
