\name{capital}
\alias{capital}
\title{Quarterly capital expenditure and appropriations}
\description{Seasonally adjusted quarterly capital expenditure and appropriations
in U.S. manufacturing: 1953--1974.}
\usage{capital}
\format{Bivariate time series containing the following columns:
\describe{\item{capital}{Quarterly capital expenditure for US manufacturing.}
\item{appropriations}{Quarterly capital appropriations for US manufacturing.}
}}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 8.}

\keyword{datasets}
\examples{plot(capital)
}
