\name{cement}
\alias{cement}
\title{Cement composition and heat data}
\description{Cement composition and heat data.}
\usage{cement}
\format{Data frame containing the following columns:
\describe{\item{pc1}{Percentage by weight of component 1}
\item{pc2}{Percentage by weight of component 2}
\item{pc3}{Percentage by weight of component 3}
\item{heat}{Heat emitted in calories per gram of cement.}
}}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 6.4}

\keyword{datasets}
\examples{plot(cement)
}
