\name{chicken}
\alias{chicken}
\title{Price of chicken}
\description{Price of chicken in US (constant dollars): 1924--1993.}
\usage{chicken}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 9.}

\keyword{datasets}
\examples{plot(chicken)
}
