\name{condmilk}
\alias{condmilk}
\title{Condensed milk}
\description{Manufacturer's Stocks of evaporated and sweetened condensed milk.}
\usage{condmilk}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 7.5.}

\keyword{datasets}
\examples{plot(condmilk)
seasonplot(condmilk)
tsdisplay(condmilk)
}
