\name{copper}
\alias{copper}
\title{Copper price}
\description{Yearly copper prices, 1800--1997  (in constant 1997 dollars).}
\usage{copper}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 9.}

\keyword{datasets}
\examples{plot(copper)
}
