\name{copper1}
\alias{copper1}
\title{Copper prices}
\description{Monthly copper prices for 28 consecutive months  (in constant 1997 dollars).}
\usage{copper1}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 9.}

\keyword{datasets}
\examples{plot(copper1)
}
