\name{copper2}
\alias{copper2}
\title{Copper prices}
\description{Yearly copper prices for 14 consecutive years (in constant 1997 dollars).}
\usage{copper2}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 9.}

\keyword{datasets}
\examples{plot(copper2)
}
