\name{copper3}
\alias{copper3}
\title{Copper prices}
\description{Yearly copper prices for 43 consecutive years (in constant 1997 dollars).}
\usage{copper3}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 9.}

\keyword{datasets}
\examples{plot(copper3)
}
