\name{Create.FMA}
\alias{Create.FMA}
\title{Create.FMA}
\description{
  Create list for frequensti model averaging estimators.
}
\usage{
Create.FMA(w, name, allMods)
}
\arguments{
  \item{w}{The weight vector.}
  \item{name}{The name of the method.}
  \item{allMods}{The resulting list from \code{\link{EstAllModels}}.}
}
\keyword{Create.FMA}