\name{dj}
\alias{dj}
\title{Dow-Jones index}
\description{Dow-Jones index on 251 trading days ending 26 Aug 1994.}
\usage{dj}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 7.}

\references{Brockwell and Davis (1996)}
\keyword{datasets}
\examples{tsdisplay(dj)
}
