\name{dole}
\alias{dole}
\title{Unemployment benefits in Australia}
\description{Monthly total of people on unemployment benefits in Australia (Jan 1965 -- Jul 1992).}
\usage{dole}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 2.3.}

\keyword{datasets}
\examples{plot(dole)
tsdisplay(dole)
}
