\name{dowjones}
\alias{dowjones}
\title{Dow-Jones index}
\description{Dow-Jones index, 28 Aug - 18 Dec 1972.}
\usage{dowjones}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 2.7.}

\keyword{datasets}
\examples{tsdisplay(dowjones)
}
