\name{eggs}
\alias{eggs}
\title{Price of eggs}
\description{Price of dozen eggs in US, 1900--1993, in constant dollars.}
\usage{eggs}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 9.}

\keyword{datasets}
\examples{plot(eggs)
}
