\name{eknives}
\alias{eknives}
\title{Sales of electric knives}
\description{Sales of electric knives: Jan 1991 - April 1992.}
\usage{eknives}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 4.2.}

\keyword{datasets}
\examples{plot(eknives)
}
