\name{elco}
\alias{elco}
\title{Sales of Elco's laser printers}
\description{Sales of Elco's laser printers: 1992--1998.}
\usage{elco}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 10.}

\keyword{datasets}
\examples{plot(elco)
}
