\name{elec}
\alias{elec}
\title{Electricity production}
\description{Australian monthly electricity production: Jan 1956 -- Aug 1995.}
\usage{elec}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapters 1--2, 7.}

\keyword{datasets}
\examples{plot(elec)
seasonplot(elec)
tsdisplay(elec)
}
