\name{EstAllModels}
\alias{EstAllModels}
\title{Estimate all submodels}
\description{
  This function estimates all submodels.
}
\usage{
EstAllModels(X, Xnew, y, S.matrix)
}
\arguments{
  \item{X}{A design matrix.}
  \item{Xnew}{Design matrix used for predictions.}
  \item{y}{Vector of the response.}
  \item{S.matrix}{Matrix where the rows correspond to a submodel.}
}
\keyword{EstAllModels}