\name{expenditure}
\alias{expenditure}
\title{Expenditure}
\description{Expenditure for 12 supermarket customers.}
\usage{expenditure}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 2.}

\keyword{datasets}
\examples{hist(expenditure)
}
