\name{fancy}
\alias{fancy}
\title{Sales for a souvenir shop}
\description{Monthly sales for a souvenir shop on the wharf at a beach resort town
in Queensland, Australia.}
\usage{fancy}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 5.8.}

\keyword{datasets}
\examples{plot(fancy)
seasonplot(fancy)
}
