\name{fma-package}
\alias{fma-package}
\alias{fma}
\docType{package}
\title{
Data sets from "Forecasting: methods and applications" by Makridakis, Wheelwright and Hyndman (1998)
}
\description{All data sets from "Forecasting: methods and applications" by
             Makridakis, Wheelwright and Hyndman (Wiley, 3rd ed., 1998).
}

\author{Rob J Hyndman. \email{Rob.Hyndman@buseco.monash.edu.au}}

\references{
Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley
& Sons: New York. \url{http://www.robhyndman.info/forecasting}}

\keyword{ package }


