% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fma.R
\name{fma}
\alias{fma}
\title{JMA}
\usage{
fma(
  y,
  x,
  ma.method = c("JMA", "MMA"),
  model.subset = c("nested", "all"),
  factor.variables = NULL,
  pd = T,
  calc.var = c("none", "boot"),
  bsa = 100
)
}
\arguments{
\item{y}{Y}

\item{x}{X}

\item{ma.method}{JMA, MMA}

\item{model.subset}{nested,all}

\item{factor.variables}{NULL}

\item{pd}{T}

\item{calc.var}{none,boot}

\item{bsa}{100}
}
\value{
betahat,se, lci , uci , weight , yhatehat , y , x ,coef of all model,Waic ,betahataic,Wbic ,betahatbic
}
\description{
JMA
}
