\name{french}
\alias{french}
\title{Industry index}
\description{French index of industry.}
\usage{french}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 4.4.}

\keyword{datasets}
\examples{plot(french)
}
