\name{hsales2}
\alias{hsales2}
\title{Sales of new one-family houses}
\description{Sales of new one-family houses in the USA (Jan 1987 -- Nov 1995).}
\usage{hsales2}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 7.10.}

\keyword{datasets}
\examples{plot(hsales2)
seasonplot(hsales2)
tsdisplay(hsales2)
}
