\name{huron}
\alias{huron}
\title{Level of Lake Huron}
\description{Level of Lake Huron in feet (reduced by 570 feet): 1875--1972.}
\usage{huron}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 8.2.}

\keyword{datasets}
\examples{plot(huron)
}
