\name{ibmclose}
\alias{ibmclose}
\title{Closing IBM stock price}
\description{Daily closing IBM stock price.}
\usage{ibmclose}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 7.2.}

\references{Box, Jenkins and Reinsell (1994) \emph{Time series
analysis: forecasting and control}, 3rd edition, Holden-Day: San
Francisco.}

\keyword{datasets}
\examples{tsdisplay(ibmclose)
}
