\name{input}
\alias{input}
\title{Input series}
\description{Input series for exercise 8.6.}
\usage{input}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 8.6.}

\keyword{datasets}
\examples{plot(input)
}
