\name{internet}
\alias{internet}
\title{Number of internet users}
\description{Number of users logged on to an internet server each minute over a 100-minute period.}
\usage{internet}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 7.}

\keyword{datasets}
\examples{tsdisplay(internet)
}
