\name{invent15}
\alias{invent15}
\title{Inventory demand}
\description{Inventory demand for product E15.}
\usage{invent15}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 2.6. Also Chapter 4.}

\keyword{datasets}
\examples{plot(invent15)
}
