\name{kkong}
\alias{kkong}
\title{King Kong data}
\description{King Kong data.}
\usage{kkong}
\format{Data frame consisting of following columns
\describe{\item{weight}{Weights of 21 gorillas}
\item{height}{Heights of 21 gorillas}
}}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 5. Exercise 5.6.}

\keyword{datasets}
\examples{plot(weight~height,data=kkong,pch=19,col=2)
}
