\name{lynx}
\alias{lynx}
\title{Annual Canadian Lynx trappings 1821--1934}
\description{Annual number of lynx trapped in McKenzie river district of northwest Canada: 1821--1934.}
\usage{lynx}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 2.3.}
\references{Campbell, M. J.and A. M. Walker (1977).
A Survey of statistical work on the Mackenzie River series of annual Canadian lynx trappings for the years 1821--1934 and a new analysis.
\emph{Journal of the Royal Statistical Society series A}, \bold{140}, 411--431.}

\keyword{datasets}
\examples{plot(lynx)
tsdisplay(lynx)
}
