\name{mink}
\alias{mink}
\title{Number of minks trapped}
\description{Annual number of minks trapped in McKenzie river district of northwest Canada: 1848--1911.}
\usage{mink}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 2.4.}

\keyword{datasets}
\examples{tsdisplay(mink)
}
