\name{nail}
\alias{nail}
\title{Nail prices}
\description{Nail prices, 1800--1996 in constant dollars.}
\usage{nail}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 9.}

\keyword{datasets}
\examples{plot(nail)
}
