\name{oilprice}
\alias{oilprice}
\title{Oil prices}
\description{Oil prices in constant 1997 dollars: 1870--1997.}
\usage{oilprice}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 10.}

\keyword{datasets}
\examples{plot(oilprice)
}
