\name{paris}
\alias{paris}
\title{Average temperature}
\description{Average monthly temperature in Paris.}
\usage{paris}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 2.1.}

\keyword{datasets}
\examples{plot(paris)
seasonplot(paris)
tsdisplay(paris)
}
