\name{pigs}
\alias{pigs}
\title{Number of pigs slaughtered}
\description{Monthly total number of pigs slaughtered in Victoria, Australia (Jan 1980 -- Aug 1995).}
\usage{pigs}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 7.}

\keyword{datasets}
\examples{tsdisplay(pigs)
}
