\name{pollution}
\alias{pollution}
\title{Shipment of pollution equipment}
\description{Monthly shipments of pollution equipment (in thousands of French francs),
Jan 1986 -- Oct 1996.}
\usage{pollution}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 7.}

\keyword{datasets}
\examples{tsdisplay(pollution)
}
