\name{productC}
\alias{productC}
\title{Sales of product C}
\description{Sales of product C (a lubricant sold in large containers).}
\usage{productC}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 1.}

\keyword{datasets}
\examples{plot(productC)
}
