\name{qelec}
\alias{qelec}
\title{Electricity production}
\description{Quarterly electricity production.}
\usage{qelec}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 3.4.}

\keyword{datasets}
\examples{plot(decompose(qelec))
}
