\name{qsales}
\alias{qsales}
\title{Sales data}
\description{Quarterly sales data.}
\usage{qsales}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 4. Exercise 4.8}

\keyword{datasets}
\examples{plot(qsales)
}
