\name{qsales}
\alias{qsales}
\title{Sales data}
\description{Quarterly exports of a French company in thousands of francs.}
\usage{qsales}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 3.7 and Table 4-7.}

\keyword{datasets}
\examples{plot(qsales)
}
