\name{schizo}
\alias{schizo}
\title{Perceptual speed scores}
\description{Daily perceptual speed scores for a schizophrenic patient. The patient began receiving a powerful tranquilizer
(chlorpromzaine) on the 61st day and continued receiving the drug for the remainder of the sample period. It is expected that
this drug would reduce perceptual speed.}
\usage{schizo}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 8.8.}

\references{McCleary and Hay (1980).}

\keyword{datasets}
\examples{plot(schizo)
}
