\name{shampoo}
\alias{shampoo}
\title{Sales of shampoo}
\description{Sales of shampoo over a three year period.}
\usage{shampoo}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 3.}

\keyword{datasets}
\examples{plot(shampoo)
}
