\name{sheep}
\alias{sheep}
\title{Sheep population}
\description{Sheep population (in millions) of England and Wales: 1867--1939.}
\usage{sheep}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York. Exercise 7.6.}
\references{Kendall (1976).}

\keyword{datasets}
\examples{tsdisplay(sheep)
}
