\name{ship}
\alias{ship}
\title{Electric can opener shipments}
\description{Electric can opener shipments.}
\usage{ship}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 4. Exercise 4.6.}

\keyword{datasets}
\examples{plot(ship)
}
