\name{shipex}
\alias{shipex}
\title{Shipments}
\description{Shipments}
\usage{shipex}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 3.1}

\keyword{datasets}
\examples{plot(shipex)
}
