\name{strikes}
\alias{strikes}
\title{Number of strikes}
\description{Number of strikes in the US from 1951 to 1980.}
\usage{strikes}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 7.4}
\references{Brockwell and Davis (1991)}

\keyword{datasets}
\examples{tsdisplay(strikes)
}
