\name{telephone}
\alias{telephone}
\title{Telephone cost}
\description{Telephone cost in San Francisco, New York: 1915--1996.}
\usage{telephone}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 9.}

\keyword{datasets}
\examples{plot(telephone)
}
