\name{ukdeaths}
\alias{ukdeaths}
\title{Total deaths and serious injuries}
\description{Monthly total deaths and serious injuries on UK roads: Jan 1975 -- Dec 1984. In February 1983, new legislation came into force requiring seat belts to be worn.}
\usage{ukdeaths}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York. Chapter 8.}

\references{Harvey (1989)}
\keyword{datasets}
\examples{plot(ukdeaths)
seasonplot(ukdeaths)
tsdisplay(ukdeaths)
}
