\name{usdeaths}
\alias{usdeaths}
\title{Accidental deaths in USA}
\description{Monthly accidental deaths in USA.}
\usage{usdeaths}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercises 2.3 and 2.4.}

\keyword{datasets}
\examples{plot(usdeaths)
seasonplot(usdeaths)
tsdisplay(usdeaths)
}
