\name{uselec}
\alias{uselec}
\title{Total generation of electricity}
\description{Monthly total generation of electricity by the U.S.
electric industry (Jan 1985 - Oct 1996.}
\usage{uselec}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 7.8.}

\keyword{datasets}
\examples{plot(uselec)
seasonplot(uselec)
tsdisplay(uselec)
}
