\name{ustreas}
\alias{ustreas}
\title{Treasury bill contracts}
\description{US treasury bill contracts on the Chicago market for 100 consecutive trading days in 1981.}
\usage{ustreas}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 1.}

\keyword{datasets}
\examples{plot(ustreas)
tsdisplay(ustreas)
}
