\name{wagesuk}
\alias{wagesuk}
\title{Real daily wages}
\description{Real daily wages in pound, England: 1260--1994.}
\usage{wagesuk}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 9.}

\keyword{datasets}
\examples{plot(wagesuk)
}
