\name{wheat}
\alias{wheat}
\title{Wheat prices}
\description{Wheat prices in constant 1996 pounds: 1264--1996.}
\usage{wheat}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 9.}

\keyword{datasets}
\examples{plot(wheat)
}
