\name{wn}
\alias{wn}
\title{White noise series}
\description{White noise series.}
\usage{wn}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Exercise 7.3.}

\keyword{datasets}
\examples{tsdisplay(wn)
}
