\name{wnoise}
\alias{wnoise}
\title{White noise time series}
\description{White noise time series with 36 values.}
\usage{wnoise}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chpater 7.}

\keyword{datasets}
\examples{tsdisplay(wnoise)
}
