\name{writing}
\alias{writing}
\title{Sales of printing and writing paper}
\description{Industry sales for printing and writing paper (in thousands of French
francs): Jan 1963 -- Dec 1972.}
\usage{writing}
\format{Time series data}
\source{Makridakis, Wheelwright and Hyndman (1998) \emph{Forecasting: methods and applications}, John Wiley & Sons: New York.
Chapter 7.}

\keyword{datasets}
\examples{tsdisplay(writing)
seasonplot(writing)
}
